# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ruckus Conference Program Operations API
  version: 1.0.0
  summary: Conference program operations - CFP, review, speakers, schedule, comms, CRM and public program feeds.
  description: >-
    HTTP API for Ruckus, an open-source conference program-operations tool. Paths are
    generated from the live Hono route registrations in src/app.ts and the src route
    modules, so path and method coverage is exact. Request and response bodies are
    intentionally loose object schemas with descriptions: handlers accept and return
    JSON envelopes shaped as a data property, and the demo does not enforce strict
    field level typing. Identity is demo persona simulation, not authentication.
  license:
    name: MIT
servers:
  - url: https://ruckus.to
    description: Live demo deployment
  - url: http://localhost:8787
    description: Local development API
...
security: []
paths:
...
  "/e/{slug}/public":
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        description: "slug path parameter."
    get:
      tags: [public-widgets]
      operationId: get_e_slug_public
      summary: "Registered in publicSite.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: true
      description: Standard failure envelope returned by the shared error helper.
      properties:
        error:
          type: object
          additionalProperties: true
          properties:
            code:
              type: string
              description: Short machine readable code such as NOT_FOUND or VALIDATION_ERROR.
            message:
              type: string
              description: Human readable explanation.